Fixed backup copy removal in try_module_insert and fixed paths in patch-realsense-ubuntu-xenial.sh

+++ scripts/patch-utils.sh
@@ -125,7 +125,10 @@ function try_module_insert {
 	else
 		# Everything went OK, delete backup
 		printf "\e[32m succeeded\n\e[0m"
-		sudo rm ${tgt_ko}.bckup
+		if [ ${backup_available} -ne 0 ];
+		then
+			sudo rm ${tgt_ko}.bckup
+		fi
 	fi
 
 	# Reload all dependent modules recursively
